% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shoots.R
\name{add_shoots}
\alias{add_shoots}
\title{Add shoots}
\usage{
add_shoots(grammar, new_root, ...)
}
\arguments{
\item{grammar}{A data table containing a Tracery-like grammar}

\item{new_root}{A (new) root string}

\item{...}{A series of shoot strings to add to the specified root}
}
\value{
A data table containing a Tracery-like grammar
}
\description{
Append new shoots to a new or existing root in a grammar.
}
