% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_grammar.R
\name{flatten_grammar}
\alias{flatten_grammar}
\title{Flatten Grammar}
\usage{
flatten_grammar(grammar, origin)
}
\arguments{
\item{grammar}{A data table containing a Tracery-like grammar}

\item{origin}{A root string to begin flattening}
}
\value{
A data table containing a Tracery-like grammar
}
\description{
Generate text from a specified origin and grammar.
}
