% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_branch.R
\name{grow_branch}
\alias{grow_branch}
\title{Grow Branch}
\usage{
grow_branch(grammar, a_root)
}
\arguments{
\item{grammar}{A data table containing a Tracery-like grammar}

\item{a_root}{A root string to seed a shoot}
}
\value{
A string
}
\description{
Generate text by growing shoots from a specified root.
}
