% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grammar.R
\name{read_grammar}
\alias{read_grammar}
\title{Read Grammar}
\usage{
read_grammar(filename)
}
\arguments{
\item{filename}{A CSV containing a Tracery-like grammar}
}
\value{
A data table containing a Tracery-like grammar
}
\description{
Read in a CSV contianing a grammar.
}
