% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_shoots.R
\name{trim_shoots}
\alias{trim_shoots}
\title{Trim shoots}
\usage{
trim_shoots(grammar, trimmed_root, ...)
}
\arguments{
\item{grammar}{A data table containing a Tracery-like grammar}

\item{trimmed_root}{Root string to remove}

\item{...}{A series of shoot strings to remove from the specified root}
}
\value{
A data table containing a Tracery-like grammar
}
\description{
Remove shoots or an entire root from a grammar.
}
